add_to() -> rte.add_to()
