+++ pyrte_rrtmgp/tests/test_cld_sw_solver.py
@@ -82,7 +82,7 @@ def test_sw_solver_with_clouds() -> None:
     optical_props["surface_albedo_diffuse"] = 0.06
     optical_props["mu0"] = 0.86
 
-    fluxes = clouds_optical_props.add_to(optical_props, delta_scale=True).rte.solve(
+    fluxes = clouds_optical_props.rte.add_to(optical_props, delta_scale=True).rte.solve(
                        add_to_input=False)
     assert fluxes is not None
 
@@ -147,7 +147,7 @@ def test_sw_solver_with_clouds_dask() -> None:
     optical_props["surface_albedo_diffuse"] = 0.06
     optical_props["mu0"] = 0.86
 
-    fluxes = clouds_optical_props.add_to(optical_props, delta_scale=True).rte.solve(
+    fluxes = clouds_optical_props.rte.add_to(optical_props, delta_scale=True).rte.solve(
                        add_to_input=False)
     assert fluxes is not None
